+++ src/api/ext-match.ts
@@ -1,8 +1,7 @@
 import { extname } from "path";
 
 function extensionMatch(filename: string, extensions: string[]): boolean {
-  const fileExt = extname(filename);
-  return extensions.map((ext) => (`.${ext}`).toLowerCase()).indexOf(fileExt.toLowerCase()) >= 0;
+  return extensions.indexOf(extname(filename).toLowerCase()) >= 0;
 }
 
 export default extensionMatch;